feat: split up large files
