+++ zork/actions.go
@@ -2,322 +2,6 @@ package zork
 
 import "math/rand"
 
-// ================================================================
-// COMBAT SYSTEM TYPES AND TABLES
-// ================================================================
-
-type BlowRes int
-
-const (
-	BlowUnk BlowRes = iota
-	BlowMissed
-	BlowUncon
-	BlowKill
-	BlowLightWnd
-	BlowHeavyWnd
-	BlowStag
-	BlowLoseWpn
-	BlowHesitate
-	BlowSitDuck
-)
-
-// Combat mode constants
-const (
-	FBusy       = 1
-	FDead       = 2
-	FUnconscious = 3
-	FConscious  = 4
-	FFirst      = 5
-)
-
-// Combat strength constants
-const (
-	StrengthMax = 7
-	StrengthMin = 2
-	CureWait    = 30
-)
-
-// Villain table field indices
-const (
-	VVillain  = 0
-	VBest     = 1
-	VBestAdv  = 2
-	VProb     = 3
-	VMsgs     = 4
-)
-
-// Melee message marker constants
-const (
-	FWep = 0 // means print weapon name
-	FDef = 1 // means print defender name
-)
-
-// MeleePart is a fragment of a melee message (either a string or FWep/FDef marker)
-type MeleePart struct {
-	Text   string
-	Marker int // -1 = normal text, FWep = weapon, FDef = defender
-}
-
-func mp(s string) MeleePart { return MeleePart{Text: s, Marker: -1} }
-func mw() MeleePart         { return MeleePart{Marker: FWep} }
-func md() MeleePart         { return MeleePart{Marker: FDef} }
-
-// A MeleeMsg is a sequence of parts that combine to form one message
-type MeleeMsg []MeleePart
-
-// MeleeSet is a set of alternative messages for one outcome type
-type MeleeSet []MeleeMsg
-
-// MeleeTable is indexed by BlowRes-1 (Missed=0, Unconscious=1, Killed=2, LightWound=3, HeavyWound=4, Stagger=5, LoseWeapon=6, Hesitate=7, SittingDuck=8)
-type MeleeTable [9]MeleeSet
-
-// VillainEntry holds a villain's combat parameters
-type VillainEntry struct {
-	Villain *Object
-	Best    *Object // best weapon against this villain
-	BestAdv int     // advantage conferred by best weapon
-	Prob    int     // probability of waking if unconscious
-	Msgs    *MeleeTable
-}
-
-// Tables of melee results
-var (
-	Def1    = [13]BlowRes{BlowMissed, BlowMissed, BlowMissed, BlowMissed, BlowStag, BlowStag, BlowUncon, BlowUncon, BlowKill, BlowKill, BlowKill, BlowKill, BlowKill}
-	Def2A   = [10]BlowRes{BlowMissed, BlowMissed, BlowMissed, BlowMissed, BlowMissed, BlowStag, BlowStag, BlowLightWnd, BlowLightWnd, BlowUncon}
-	Def2B   = [12]BlowRes{BlowMissed, BlowMissed, BlowMissed, BlowStag, BlowStag, BlowLightWnd, BlowLightWnd, BlowLightWnd, BlowUncon, BlowKill, BlowKill, BlowKill}
-	Def3A   = [11]BlowRes{BlowMissed, BlowMissed, BlowMissed, BlowMissed, BlowMissed, BlowStag, BlowStag, BlowLightWnd, BlowLightWnd, BlowHeavyWnd, BlowHeavyWnd}
-	Def3B   = [11]BlowRes{BlowMissed, BlowMissed, BlowMissed, BlowStag, BlowStag, BlowLightWnd, BlowLightWnd, BlowLightWnd, BlowHeavyWnd, BlowHeavyWnd, BlowHeavyWnd}
-	Def3C   = [10]BlowRes{BlowMissed, BlowStag, BlowStag, BlowLightWnd, BlowLightWnd, BlowLightWnd, BlowLightWnd, BlowHeavyWnd, BlowHeavyWnd, BlowHeavyWnd}
-	Def1Res = [4]BlowRes{Def1[0], BlowUnk, BlowUnk}
-	Def2Res = [4]BlowRes{Def2A[0], Def2B[0], BlowUnk, BlowUnk}
-	Def3Res = [5]BlowRes{Def3A[0], BlowUnk, Def3B[0], BlowUnk, Def3C[0]}
-
-	// Hero melee messages (indexed by outcome: missed, unconscious, killed, light-wound, heavy-wound, stagger, lose-weapon)
-	HeroMelee = MeleeTable{
-		// BlowMissed
-		{
-			{mp("Your "), mw(), mp(" misses the "), md(), mp(" by an inch.")},
-			{mp("A good slash, but it misses the "), md(), mp(" by a mile.")},
-			{mp("You charge, but the "), md(), mp(" jumps nimbly aside.")},
-			{mp("Clang! Crash! The "), md(), mp(" parries.")},
-			{mp("A quick stroke, but the "), md(), mp(" is on guard.")},
-			{mp("A good stroke, but it's too slow; the "), md(), mp(" dodges.")},
-		},
-		// BlowUncon
-		{
-			{mp("Your "), mw(), mp(" crashes down, knocking the "), md(), mp(" into dreamland.")},
-			{mp("The "), md(), mp(" is battered into unconsciousness.")},
-			{mp("A furious exchange, and the "), md(), mp(" is knocked out!")},
-			{mp("The haft of your "), mw(), mp(" knocks out the "), md(), mp(".")},
-			{mp("The "), md(), mp(" is knocked out!")},
-		},
-		// BlowKill
-		{
-			{mp("It's curtains for the "), md(), mp(" as your "), mw(), mp(" removes his head.")},
-			{mp("The fatal blow strikes the "), md(), mp(" square in the heart: He dies.")},
-			{mp("The "), md(), mp(" takes a fatal blow and slumps to the floor dead.")},
-		},
-		// BlowLightWnd
-		{
-			{mp("The "), md(), mp(" is struck on the arm; blood begins to trickle down.")},
-			{mp("Your "), mw(), mp(" pinks the "), md(), mp(" on the wrist, but it's not serious.")},
-			{mp("Your stroke lands, but it was only the flat of the blade.")},
-			{mp("The blow lands, making a shallow gash in the "), md(), mp("'s arm!")},
-		},
-		// BlowHeavyWnd
-		{
-			{mp("The "), md(), mp(" receives a deep gash in his side.")},
-			{mp("A savage blow on the thigh! The "), md(), mp(" is stunned but can still fight!")},
-			{mp("Slash! Your blow lands! That one hit an artery, it could be serious!")},
-			{mp("Slash! Your stroke connects! This could be serious!")},
-		},
-		// BlowStag
-		{
-			{mp("The "), md(), mp(" is staggered, and drops to his knees.")},
-			{mp("The "), md(), mp(" is momentarily disoriented and can't fight back.")},
-			{mp("The force of your blow knocks the "), md(), mp(" back, stunned.")},
-			{mp("The "), md(), mp(" is confused and can't fight back.")},
-			{mp("The quickness of your thrust knocks the "), md(), mp(" back, stunned.")},
-		},
-		// BlowLoseWpn
-		{
-			{mp("The "), md(), mp("'s weapon is knocked to the floor, leaving him unarmed.")},
-			{mp("The "), md(), mp(" is disarmed by a subtle feint past his guard.")},
-		},
-		// BlowHesitate (not used for hero)
-		{},
-		// BlowSitDuck (not used for hero)
-		{},
-	}
-
-	// Cyclops melee messages
-	CyclopsMelee = MeleeTable{
-		// BlowMissed
-		{
-			{mp("The Cyclops misses, but the backwash almost knocks you over.")},
-			{mp("The Cyclops rushes you, but runs into the wall.")},
-		},
-		// BlowUncon
-		{
-			{mp("The Cyclops sends you crashing to the floor, unconscious.")},
-		},
-		// BlowKill
-		{
-			{mp("The Cyclops breaks your neck with a massive smash.")},
-		},
-		// BlowLightWnd
-		{
-			{mp("A quick punch, but it was only a glancing blow.")},
-			{mp("A glancing blow from the Cyclops' fist.")},
-		},
-		// BlowHeavyWnd
-		{
-			{mp("The monster smashes his huge fist into your chest, breaking several ribs.")},
-			{mp("The Cyclops almost knocks the wind out of you with a quick punch.")},
-		},
-		// BlowStag
-		{
-			{mp("The Cyclops lands a punch that knocks the wind out of you.")},
-			{mp("Heedless of your weapons, the Cyclops tosses you against the rock wall of the room.")},
-		},
-		// BlowLoseWpn
-		{
-			{mp("The Cyclops grabs your "), mw(), mp(", tastes it, and throws it to the ground in disgust.")},
-			{mp("The monster grabs you on the wrist, squeezes, and you drop your "), mw(), mp(" in pain.")},
-		},
-		// BlowHesitate
-		{
-			{mp("The Cyclops seems unable to decide whether to broil or stew his dinner.")},
-		},
-		// BlowSitDuck
-		{
-			{mp("The Cyclops, no sportsman, dispatches his unconscious victim.")},
-		},
-	}
-
-	// Troll melee messages
-	TrollMelee = MeleeTable{
-		// BlowMissed
-		{
-			{mp("The troll swings his axe, but it misses.")},
-			{mp("The troll's axe barely misses your ear.")},
-			{mp("The axe sweeps past as you jump aside.")},
-			{mp("The axe crashes against the rock, throwing sparks!")},
-		},
-		// BlowUncon
-		{
-			{mp("The flat of the troll's axe hits you delicately on the head, knocking you out.")},
-		},
-		// BlowKill
-		{
-			{mp("The troll neatly removes your head.")},
-			{mp("The troll's axe stroke cleaves you from the nave to the chops.")},
-			{mp("The troll's axe removes your head.")},
-		},
-		// BlowLightWnd
-		{
-			{mp("The axe gets you right in the side. Ouch!")},
-			{mp("The flat of the troll's axe skins across your forearm.")},
-			{mp("The troll's swing almost knocks you over as you barely parry in time.")},
-			{mp("The troll swings his axe, and it nicks your arm as you dodge.")},
-		},
-		// BlowHeavyWnd
-		{
-			{mp("The troll charges, and his axe slashes you on your "), mw(), mp(" arm.")},
-			{mp("An axe stroke makes a deep wound in your leg.")},
-			{mp("The troll's axe swings down, gashing your shoulder.")},
-		},
-		// BlowStag
-		{
-			{mp("The troll hits you with a glancing blow, and you are momentarily stunned.")},
-			{mp("The troll swings; the blade turns on your armor but crashes broadside into your head.")},
-			{mp("You stagger back under a hail of axe strokes.")},
-			{mp("The troll's mighty blow drops you to your knees.")},
-		},
-		// BlowLoseWpn
-		{
-			{mp("The axe hits your "), mw(), mp(" and knocks it spinning.")},
-			{mp("The troll swings, you parry, but the force of his blow knocks your "), mw(), mp(" away.")},
-			{mp("The axe knocks your "), mw(), mp(" out of your hand. It falls to the floor.")},
-		},
-		// BlowHesitate
-		{
-			{mp("The troll hesitates, fingering his axe.")},
-			{mp("The troll scratches his head ruminatively:  Might you be magically protected, he wonders?")},
-		},
-		// BlowSitDuck
-		{
-			{mp("Conquering his fears, the troll puts you to death.")},
-		},
-	}
-
-	// Thief melee messages
-	ThiefMelee = MeleeTable{
-		// BlowMissed
-		{
-			{mp("The thief stabs nonchalantly with his stiletto and misses.")},
-			{mp("You dodge as the thief comes in low.")},
-			{mp("You parry a lightning thrust, and the thief salutes you with a grim nod.")},
-			{mp("The thief tries to sneak past your guard, but you twist away.")},
-		},
-		// BlowUncon
-		{
-			{mp("Shifting in the midst of a thrust, the thief knocks you unconscious with the haft of his stiletto.")},
-			{mp("The thief knocks you out.")},
-		},
-		// BlowKill
-		{
-			{mp("Finishing you off, the thief inserts his blade into your heart.")},
-			{mp("The thief comes in from the side, feints, and inserts the blade into your ribs.")},
-			{mp("The thief bows formally, raises his stiletto, and with a wry grin, ends the battle and your life.")},
-		},
-		// BlowLightWnd
-		{
-			{mp("A quick thrust pinks your left arm, and blood starts to trickle down.")},
-			{mp("The thief draws blood, raking his stiletto across your arm.")},
-			{mp("The stiletto flashes faster than you can follow, and blood wells from your leg.")},
-			{mp("The thief slowly approaches, strikes like a snake, and leaves you wounded.")},
-		},
-		// BlowHeavyWnd
-		{
-			{mp("The thief strikes like a snake! The resulting wound is serious.")},
-			{mp("The thief stabs a deep cut in your upper arm.")},
-			{mp("The stiletto touches your forehead, and the blood obscures your vision.")},
-			{mp("The thief strikes at your wrist, and suddenly your grip is slippery with blood.")},
-		},
-		// BlowStag
-		{
-			{mp("The butt of his stiletto cracks you on the skull, and you stagger back.")},
-			{mp("The thief rams the haft of his blade into your stomach, leaving you out of breath.")},
-			{mp("The thief attacks, and you fall back desperately.")},
-		},
-		// BlowLoseWpn
-		{
-			{mp("A long, theatrical slash. You catch it on your "), mw(), mp(", but the thief twists his knife, and the "), mw(), mp(" goes flying.")},
-			{mp("The thief neatly flips your "), mw(), mp(" out of your hands, and it drops to the floor.")},
-			{mp("You parry a low thrust, and your "), mw(), mp(" slips out of your hand.")},
-		},
-		// BlowHesitate
-		{
-			{mp("The thief, a man of superior breeding, pauses for a moment to consider the propriety of finishing you off.")},
-			{mp("The thief amuses himself by searching your pockets.")},
-			{mp("The thief entertains himself by rifling your pack.")},
-		},
-		// BlowSitDuck
-		{
-			{mp("The thief, forgetting his essentially genteel upbringing, cuts your throat.")},
-			{mp("The thief, a pragmatist, dispatches you as a threat to his livelihood.")},
-		},
-	}
-
-	// Villain table - initialized in FinalizeGameObjects to avoid init cycles
-	Villains []*VillainEntry
-)
-
-// ================================================================
-// HELPER FUNCTIONS
-// ================================================================
-
 func OpenClose(obj *Object, strOpn, strCls string) {
 	if ActVerb.Norm == "open" {
 		if obj.Has(FlgOpen) {
@@ -334,47 +18,6 @@ func OpenClose(obj *Object, strOpn, strCls string) {
 			Print(PickOne(Dummy), Newline)
 		}
 	}
-}
-
-func Rob(what, where *Object, prob int) bool {
-	robbed := false
-	for _, x := range what.Children {
-		if x.Has(FlgInvis) || x.Has(FlgSacred) {
-			continue
-		}
-		if x.TValue <= 0 {
-			continue
-		}
-		if prob > 0 && !Prob(prob, false) {
-			continue
-		}
-		x.MoveTo(where)
-		x.Give(FlgTouch)
-		if where == &Thief {
-			x.Give(FlgInvis)
-		}
-		robbed = true
-	}
-	return robbed
-}
-
-func WeaponFunction(w, v *Object) bool {
-	if !v.IsIn(Here) {
-		return false
-	}
-	if ActVerb.Norm == "take" {
-		if w.IsIn(v) {
-			Print("The ", NoNewline)
-			PrintObject(v)
-			Print(" swings it out of your reach.", Newline)
-		} else {
-			Print("The ", NoNewline)
-			PrintObject(w)
-			Print(" seems white-hot. You can't hold on to it.", Newline)
-		}
-		return true
-	}
-	return false
 }
 
 func LeavesAppear() bool {
@@ -439,18 +82,6 @@ func WithTell(obj *Object) {
 	Print("?", Newline)
 }
 
-func FixBoat() {
-	Print("Well done. The boat is repaired.", Newline)
-	InflatableBoat.MoveTo(PuncturedBoat.Location())
-	RemoveCarefully(&PuncturedBoat)
-}
-
-
-func FixMaintLeak() {
-	WaterLevel = -1
-	QueueInt(IMaintRoom, false).Run = false
-	Print("By some miracle of Zorkian technology, you have managed to stop the leak in the dam.", Newline)
-}
 
 func BadEgg() {
 	if Canary.IsIn(&Egg) {
@@ -483,102 +114,7 @@ func ForestRoomQ() bool {
 		Here == &Path || Here == &UpATree
 }
 
-func StolenLight() bool {
-	oLit := Lit
-	Lit = IsLit(Here, true)
-	if !Lit && oLit {
-		Print("The thief seems to have left you in the dark.", Newline)
-	}
-	return true
-}
-
-func RecoverStiletto() {
-	if Stiletto.IsIn(Thief.Location()) {
-		Stiletto.Give(FlgNoDesc)
-		Stiletto.MoveTo(&Thief)
-	}
-}
-
-func HackTreasures() {
-	RecoverStiletto()
-	Thief.Give(FlgInvis)
-	for _, x := range TreasureRoom.Children {
-		x.Take(FlgInvis)
-	}
-}
-
-func DepositBooty(rm *Object) bool {
-	flg := false
-	var toMove []*Object
-	for _, x := range Thief.Children {
-		if x == &Stiletto || x == &LargeBag {
-			continue
-		}
-		if x.TValue > 0 {
-			toMove = append(toMove, x)
-			flg = true
-			if x == &Egg {
-				EggSolve = true
-				Egg.Give(FlgOpen)
-			}
-		}
-	}
-	for _, x := range toMove {
-		x.MoveTo(rm)
-	}
-	return flg
-}
-
-func RobMaze(rm *Object) bool {
-	for _, x := range rm.Children {
-		if x.Has(FlgTake) && !x.Has(FlgInvis) && Prob(40, false) {
-			Print("You hear, off in the distance, someone saying \"My, I wonder what this fine ", NoNewline)
-			PrintObject(x)
-			Print(" is doing here.\"", Newline)
-			if Prob(60, true) {
-				x.MoveTo(&Thief)
-				x.Give(FlgTouch)
-				x.Give(FlgInvis)
-			}
-			return true
-		}
-	}
-	return false
-}
-
 // StealJunk and DropJunk are defined in the IThief section below
-
-func MoveAll(from, to *Object) {
-	var toMove []*Object
-	for _, x := range from.Children {
-		toMove = append(toMove, x)
-	}
-	for _, x := range toMove {
-		x.Take(FlgInvis)
-		x.MoveTo(to)
-	}
-}
-
-func ThiefInTreasure() {
-	if len(Here.Children) > 1 {
-		Print("The thief gestures mysteriously, and the treasures in the room suddenly vanish.", Newline)
-		NewLine()
-	}
-	for _, f := range Here.Children {
-		if f != &Chalice && f != &Thief {
-			f.Give(FlgInvis)
-		}
-	}
-}
-
-func Infested(r *Object) bool {
-	for _, f := range r.Children {
-		if f.Has(FlgActor) && !f.Has(FlgInvis) {
-			return true
-		}
-	}
-	return false
-}
 
 func MinInt(a, b int) int {
 	if a < b {
@@ -592,44 +128,6 @@ func AbsInt(a int) int {
 		return -a
 	}
 	return a
-}
-
-func FightStrength(adjust bool) int {
-	if ScoreMax == 0 {
-		return StrengthMin
-	}
-	s := StrengthMin + Score/(ScoreMax/(StrengthMax-StrengthMin))
-	if adjust {
-		s += Winner.Strength
-	}
-	return s
-}
-
-func FindWeapon(o *Object) *Object {
-	for _, w := range o.Children {
-		if w == &Stiletto || w == &Axe || w == &Sword || w == &Knife || w == &RustyKnife {
-			return w
-		}
-	}
-	return nil
-}
-
-func Winning(v *Object) bool {
-	vs := v.Strength
-	ps := vs - FightStrength(true)
-	if ps > 3 {
-		return Prob(90, false)
-	}
-	if ps > 0 {
-		return Prob(75, false)
-	}
-	if ps == 0 {
-		return Prob(50, false)
-	}
-	if vs > 1 {
-		return Prob(25, false)
-	}
-	return Prob(10, false)
 }
 
 // ================================================================
@@ -1015,68 +513,6 @@ func ForestFcn(arg ActArg) bool {
 	return false
 }
 
-func WaterFcn(arg ActArg) bool {
-	if ActVerb.Norm == "sgive" {
-		return false
-	}
-	if ActVerb.Norm == "through" || ActVerb.Norm == "board" {
-		Print(PickOne(SwimYuks), Newline)
-		return true
-	}
-	// Simplified water handling
-	if ActVerb.Norm == "take" || ActVerb.Norm == "put" {
-		w := DirObj
-		if w == &GlobalWater {
-			w = &Water
-		}
-		if ActVerb.Norm == "take" {
-			if w.IsIn(&Bottle) && IndirObj == nil {
-				Print("It's in the bottle. Perhaps you should take that instead.", Newline)
-				return true
-			}
-			if Bottle.IsIn(Winner) {
-				if !Bottle.Has(FlgOpen) {
-					Print("The bottle is closed.", Newline)
-					ThisIsIt(&Bottle)
-					return true
-				}
-				if !Bottle.HasChildren() {
-					Water.MoveTo(&Bottle)
-					Print("The bottle is now full of water.", Newline)
-					return true
-				}
-				Print("The water slips through your fingers.", Newline)
-				return true
-			}
-			Print("The water slips through your fingers.", Newline)
-			return true
-		}
-	}
-	if ActVerb.Norm == "drop" || ActVerb.Norm == "give" {
-		if ActVerb.Norm == "drop" && Water.IsIn(&Bottle) && !Bottle.Has(FlgOpen) {
-			Print("The bottle is closed.", Newline)
-			return true
-		}
-		RemoveCarefully(&Water)
-		av := Winner.Location()
-		if av.Has(FlgVeh) {
-			Print("There is now a puddle in the bottom of the ", NoNewline)
-			PrintObject(av)
-			Print(".", Newline)
-			Water.MoveTo(av)
-		} else {
-			Print("The water spills to the floor and evaporates immediately.", Newline)
-		}
-		return true
-	}
-	if ActVerb.Norm == "throw" {
-		Print("The water splashes on the walls and evaporates immediately.", Newline)
-		RemoveCarefully(&Water)
-		return true
-	}
-	return false
-}
-
 func KitchenWindowFcn(arg ActArg) bool {
 	if ActVerb.Norm == "open" || ActVerb.Norm == "close" {
 		KitchenWindowFlag = true
@@ -1248,70 +684,6 @@ func AxeFcn(arg ActArg) bool {
 		return false
 	}
 	return WeaponFunction(&Axe, &Troll)
-}
-
-func BoltFcn(arg ActArg) bool {
-	if ActVerb.Norm == "turn" {
-		if IndirObj == &Wrench {
-			if GateFlag {
-				ReservoirSouth.Take(FlgTouch)
-				if GatesOpen {
-					GatesOpen = false
-					LoudRoom.Take(FlgTouch)
-					Print("The sluice gates close and water starts to collect behind the dam.", Newline)
-					Queue(IRfill, 8).Run = true
-					QueueInt(IRempty, false).Run = false
-				} else {
-					GatesOpen = true
-					Print("The sluice gates open and water pours through the dam.", Newline)
-					Queue(IRempty, 8).Run = true
-					QueueInt(IRfill, false).Run = false
-				}
-			} else {
-				Print("The bolt won't turn with your best effort.", Newline)
-			}
-		} else {
-			Print("The bolt won't turn using the ", NoNewline)
-			PrintObject(IndirObj)
-			Print(".", Newline)
-		}
-		return true
-	}
-	if ActVerb.Norm == "take" {
-		IntegralPart()
-		return true
-	}
-	if ActVerb.Norm == "oil" {
-		Print("Hmm. It appears the tube contained glue, not oil. Turning the bolt won't get any easier....", Newline)
-		return true
-	}
-	return false
-}
-
-func BubbleFcn(arg ActArg) bool {
-	if ActVerb.Norm == "take" {
-		IntegralPart()
-		return true
-	}
-	return false
-}
-
-func DamFunction(arg ActArg) bool {
-	if ActVerb.Norm == "open" || ActVerb.Norm == "close" {
-		Print("Sounds reasonable, but this isn't how.", Newline)
-		return true
-	}
-	if ActVerb.Norm == "plug" {
-		if IndirObj == &Hands {
-			Print("Are you the little Dutch boy, then? Sorry, this is a big dam.", Newline)
-		} else {
-			Print("With a ", NoNewline)
-			PrintObject(IndirObj)
-			Print("? Do you know how big this dam is? You could only stop a tiny leak with that.", Newline)
-		}
-		return true
-	}
-	return false
 }
 
 func TrapDoorFcn(arg ActArg) bool {
@@ -1582,55 +954,6 @@ func LeafPileFcn(arg ActArg) bool {
 	}
 	if ActVerb.Norm == "look under" && !GrateRevealed {
 		Print("Underneath the pile of leaves is a grating. As you release the leaves, the grating is once again concealed from view.", Newline)
-		return true
-	}
-	return false
-}
-
-func PuncturedBoatFcn(arg ActArg) bool {
-	if (ActVerb.Norm == "put" || ActVerb.Norm == "put on") && DirObj == &Putty {
-		FixBoat()
-		return true
-	}
-	if ActVerb.Norm == "inflate" || ActVerb.Norm == "fill" {
-		Print("No chance. Some moron punctured it.", Newline)
-		return true
-	}
-	if ActVerb.Norm == "plug" {
-		if IndirObj == &Putty {
-			FixBoat()
-			return true
-		}
-		WithTell(IndirObj)
-		return true
-	}
-	return false
-}
-
-func InflatableBoatFcn(arg ActArg) bool {
-	if ActVerb.Norm == "inflate" || ActVerb.Norm == "fill" {
-		if !InflatableBoat.IsIn(Here) {
-			Print("The boat must be on the ground to be inflated.", Newline)
-			return true
-		}
-		if IndirObj == &Pump {
-			Print("The boat inflates and appears seaworthy.", Newline)
-			if !BoatLabel.Has(FlgTouch) {
-				Print("A tan label is lying inside the boat.", Newline)
-			}
-			Deflate = false
-			RemoveCarefully(&InflatableBoat)
-			InflatedBoat.MoveTo(Here)
-			ThisIsIt(&InflatedBoat)
-			return true
-		}
-		if IndirObj == &Lungs {
-			Print("You don't have enough lung power to inflate it.", Newline)
-			return true
-		}
-		Print("With a ", NoNewline)
-		PrintObject(IndirObj)
-		Print("? Surely you jest!", Newline)
 		return true
 	}
 	return false
@@ -2433,182 +1756,6 @@ func CyclopsFcn(arg ActArg) bool {
 // THIEF / ROBBER
 // ================================================================
 
-func RobberFcn(arg ActArg) bool {
-	if ActVerb.Norm == "tell" {
-		Print("The thief is a strong, silent type.", Newline)
-		Params.Continue = NumUndef
-		return true
-	}
-	if arg == ActArg(FBusy) {
-		if Stiletto.IsIn(&Thief) {
-			return false
-		}
-		if Stiletto.IsIn(Thief.Location()) {
-			Stiletto.MoveTo(&Thief)
-			Stiletto.Give(FlgNoDesc)
-			if Thief.IsIn(Here) {
-				Print("The robber, somewhat surprised at this turn of events, nimbly retrieves his stiletto.", Newline)
-			}
-			return true
-		}
-		return false
-	}
-	if arg == ActArg(FDead) {
-		Stiletto.MoveTo(Here)
-		Stiletto.Take(FlgNoDesc)
-		x := DepositBooty(Here)
-		if Here == &TreasureRoom {
-			flg := false
-			for _, obj := range Here.Children {
-				if obj == &Chalice || obj == &Thief || obj == &Adventurer {
-					continue
-				}
-				obj.Take(FlgInvis)
-				if !flg {
-					flg = true
-					Print("As the thief dies, the power of his magic decreases, and his treasures reappear:", Newline)
-				}
-				Print("  A ", NoNewline)
-				PrintObject(obj)
-				if obj.HasChildren() && CanSeeInside(obj) {
-					Print(", with ", NoNewline)
-					PrintContents(obj)
-				}
-				NewLine()
-			}
-			if !flg {
-				Print("The chalice is now safe to take.", Newline)
-			}
-		} else if x {
-			Print("His booty remains.", Newline)
-		}
-		QueueInt(IThief, false).Run = false
-		return true
-	}
-	if arg == ActArg(FFirst) {
-		if ThiefHere && !Thief.Has(FlgInvis) && Prob(20, false) {
-			Thief.Give(FlgFight)
-			Params.Continue = NumUndef
-			return true
-		}
-		return false
-	}
-	if arg == ActArg(FUnconscious) {
-		QueueInt(IThief, false).Run = false
-		Thief.Take(FlgFight)
-		Stiletto.MoveTo(Here)
-		Stiletto.Take(FlgNoDesc)
-		Thief.LongDesc = RobberUDesc
-		return true
-	}
-	if arg == ActArg(FConscious) {
-		if Thief.Location() == Here {
-			Thief.Give(FlgFight)
-			Print("The robber revives, briefly feigning continued unconsciousness, and, when he sees his moment, scrambles away from you.", Newline)
-		}
-		Queue(IThief, -1).Run = true
-		Thief.LongDesc = RobberCDesc
-		RecoverStiletto()
-		return true
-	}
-
-	// Default (no special mode)
-	if ActVerb.Norm == "hello" && Thief.LongDesc == RobberUDesc {
-		Print("The thief, being temporarily incapacitated, is unable to acknowledge your greeting with his usual graciousness.", Newline)
-		return true
-	}
-	if DirObj == &Knife && ActVerb.Norm == "throw" && !Thief.Has(FlgFight) {
-		DirObj.MoveTo(Here)
-		if Prob(10, false) {
-			Print("You evidently frightened the robber, though you didn't hit him. He flees", NoNewline)
-			LargeBag.Remove()
-			hasStiletto := false
-			if Stiletto.IsIn(&Thief) {
-				Stiletto.Remove()
-				hasStiletto = true
-			}
-			if Thief.HasChildren() {
-				MoveAll(&Thief, Here)
-				Print(", but the contents of his bag fall on the floor.", NoNewline)
-			} else {
-				Print(".", NoNewline)
-			}
-			LargeBag.MoveTo(&Thief)
-			if hasStiletto {
-				Stiletto.MoveTo(&Thief)
-			}
-			NewLine()
-			Thief.Give(FlgInvis)
-		} else {
-			Print("You missed. The thief makes no attempt to take the knife, though it would be a fine addition to the collection in his bag. He does seem angered by your attempt.", Newline)
-			Thief.Give(FlgFight)
-		}
-		return true
-	}
-	if (ActVerb.Norm == "throw" || ActVerb.Norm == "give") && DirObj != nil && DirObj != &Thief && IndirObj == &Thief {
-		if Thief.Strength < 0 {
-			Thief.Strength = -Thief.Strength
-			Queue(IThief, -1).Run = true
-			RecoverStiletto()
-			Thief.LongDesc = RobberCDesc
-			Print("Your proposed victim suddenly recovers consciousness.", Newline)
-		}
-		DirObj.MoveTo(&Thief)
-		if DirObj.TValue > 0 {
-			ThiefEngrossed = true
-			Print("The thief is taken aback by your unexpected generosity, but accepts the ", NoNewline)
-			PrintObject(DirObj)
-			Print(" and stops to admire its beauty.", Newline)
-		} else {
-			Print("The thief places the ", NoNewline)
-			PrintObject(DirObj)
-			Print(" in his bag and thanks you politely.", Newline)
-		}
-		return true
-	}
-	if ActVerb.Norm == "take" {
-		Print("Once you got him, what would you do with him?", Newline)
-		return true
-	}
-	if ActVerb.Norm == "examine" || ActVerb.Norm == "look inside" {
-		Print("The thief is a slippery character with beady eyes that flit back and forth. He carries, along with an unmistakable arrogance, a large bag over his shoulder and a vicious stiletto, whose blade is aimed menacingly in your direction. I'd watch out if I were you.", Newline)
-		return true
-	}
-	if ActVerb.Norm == "listen" {
-		Print("The thief says nothing, as you have not been formally introduced.", Newline)
-		return true
-	}
-	return false
-}
-
-func LargeBagFcn(arg ActArg) bool {
-	if ActVerb.Norm == "take" {
-		if Thief.LongDesc == RobberUDesc {
-			Print("Sadly for you, the robber collapsed on top of the bag. Trying to take it would wake him.", Newline)
-		} else {
-			Print("The bag will be taken over his dead body.", Newline)
-		}
-		return true
-	}
-	if ActVerb.Norm == "put" && IndirObj == &LargeBag {
-		Print("It would be a good trick.", Newline)
-		return true
-	}
-	if ActVerb.Norm == "open" || ActVerb.Norm == "close" {
-		Print("Getting close enough would be a good trick.", Newline)
-		return true
-	}
-	if ActVerb.Norm == "examine" || ActVerb.Norm == "look inside" {
-		Print("The bag is underneath the thief, so one can't say what, if anything, is inside.", Newline)
-		return true
-	}
-	return false
-}
-
-func StiletteFcn(arg ActArg) bool {
-	return WeaponFunction(&Stiletto, &Thief)
-}
-
 func DumbContainerFcn(arg ActArg) bool {
 	if ActVerb.Norm == "open" || ActVerb.Norm == "close" || ActVerb.Norm == "look inside" {
 		Print("You can't do that.", Newline)
@@ -2758,40 +1905,6 @@ func RainbowFcn(arg ActArg) bool {
 		Print("The Frigid River flows under the rainbow.", Newline)
 		return true
 	}
-	return false
-}
-
-func RiverFcn(arg ActArg) bool {
-	if ActVerb.Norm == "put" && IndirObj == &River {
-		if DirObj == &Me {
-			JigsUp("You splash around for a while, fighting the current, then you drown.", false)
-			return true
-		}
-		if DirObj == &InflatedBoat {
-			Print("You should get in the boat then launch it.", Newline)
-			return true
-		}
-		if DirObj.Has(FlgBurn) {
-			RemoveCarefully(DirObj)
-			Print("The ", NoNewline)
-			PrintObject(DirObj)
-			Print(" floats for a moment, then sinks.", Newline)
-			return true
-		}
-		RemoveCarefully(DirObj)
-		Print("The ", NoNewline)
-		PrintObject(DirObj)
-		Print(" splashes into the water and is gone forever.", Newline)
-		return true
-	}
-	if ActVerb.Norm == "leap" || ActVerb.Norm == "through" {
-		Print("A look before leaping reveals that the river is wide and dangerous, with swift currents and large, half-hidden rocks. You decide to forgo your swim.", Newline)
-		return true
-	}
-	return false
-}
-
-func TreasureInsideFcn(arg ActArg) bool {
 	return false
 }
 
@@ -3353,28 +2466,6 @@ func SouthTempleFcn(arg ActArg) bool {
 	return false
 }
 
-func DamRoomFcn(arg ActArg) bool {
-	if arg == ActLook {
-		Print("You are standing on the top of the Flood Control Dam #3, which was quite a tourist attraction in times far distant. There are paths to the north, south, and west, and a scramble down.", Newline)
-		if LowTide && GatesOpen {
-			Print("The water level behind the dam is low: The sluice gates have been opened. Water rushes through the dam and downstream.", Newline)
-		} else if GatesOpen {
-			Print("The sluice gates are open, and water rushes through the dam. The water level behind the dam is still high.", Newline)
-		} else if LowTide {
-			Print("The sluice gates are closed. The water level in the reservoir is quite low, but the level is rising quickly.", Newline)
-		} else {
-			Print("The sluice gates on the dam are closed. Behind the dam, there can be seen a wide reservoir. Water is pouring over the top of the now abandoned dam.", Newline)
-		}
-		Print("There is a control panel here, on which a large metal bolt is mounted. Directly above the bolt is a small green plastic bubble", NoNewline)
-		if GateFlag {
-			Print(" which is glowing serenely", NoNewline)
-		}
-		Print(".", Newline)
-		return true
-	}
-	return false
-}
-
 func MachineRoomFcn(arg ActArg) bool {
 	if arg == ActLook {
 		Print("This is a large, cold room whose sole exit is to the north. In one corner there is a machine which is reminiscent of a clothes dryer. On its face is a switch which is labelled \"START\". The switch does not appear to be manipulable by any human hand (unless the fingers are about 1/16 by 1/4 inch). On the front of the machine is a large lid, which is ", NoNewline)
@@ -3512,40 +2603,6 @@ func NoObjsFcn(arg ActArg) bool {
 	return false
 }
 
-func WhiteCliffsFcn(arg ActArg) bool {
-	if arg == ActEnd {
-		if InflatedBoat.IsIn(Winner) {
-			Deflate = false
-		} else {
-			Deflate = true
-		}
-	}
-	return false
-}
-
-func FallsRoomFcn(arg ActArg) bool {
-	if arg == ActLook {
-		Print("You are at the top of Aragain Falls, an enormous waterfall with a drop of about 450 feet. The only path here is on the north end.", Newline)
-		if RainbowFlag {
-			Print("A solid rainbow spans the falls.", Newline)
-		} else {
-			Print("A beautiful rainbow can be seen over the falls and to the west.", Newline)
-		}
-		return true
-	}
-	return false
-}
-
-func Rivr4RoomFcn(arg ActArg) bool {
-	if arg == ActEnd {
-		if Buoy.IsIn(Winner) && BuoyFlag {
-			Print("You notice something funny about the feel of the buoy.", Newline)
-			BuoyFlag = false
-		}
-	}
-	return false
-}
-
 func CanyonViewFcn(arg ActArg) bool {
 	return false
 }
@@ -3611,111 +2668,6 @@ func TreeRoomFcn(arg ActArg) bool {
 	}
 	if arg == ActEnter {
 		Queue(IForestRandom, -1).Run = true
-	}
-	return false
-}
-
-func RBoatFcn(arg ActArg) bool {
-	if arg == ActEnter || arg == ActEnd || arg == ActLook {
-		return false
-	}
-	if arg == ActBegin {
-		if ActVerb.Norm == "walk" {
-			if DirObj == ToDirObj("land") || DirObj == ToDirObj("east") || DirObj == ToDirObj("west") {
-				return false
-			}
-			if Here == &Reservoir && (DirObj == ToDirObj("north") || DirObj == ToDirObj("south")) {
-				return false
-			}
-			if Here == &InStream && DirObj == ToDirObj("south") {
-				return false
-			}
-			Print("Read the label for the boat's instructions.", Newline)
-			return true
-		}
-		if ActVerb.Norm == "launch" {
-			if Here == &River1 || Here == &River2 || Here == &River3 || Here == &River4 || Here == &Reservoir || Here == &InStream {
-				Print("You are on the ", NoNewline)
-				if Here == &Reservoir {
-					Print("reservoir", NoNewline)
-				} else if Here == &InStream {
-					Print("stream", NoNewline)
-				} else {
-					Print("river", NoNewline)
-				}
-				Print(", or have you forgotten?", Newline)
-				return true
-			}
-			tmp := GoNext(RiverLaunch)
-			if tmp == 1 {
-				// ZIL: <ENABLE <QUEUE I-RIVER <LKP ,HERE ,RIVER-SPEEDS>>>
-				// After GoNext, HERE is the destination room. Use its speed.
-				if spd, ok := RiverSpeedMap[Here]; ok {
-					Queue(IRiver, spd).Run = true
-				}
-				return true
-			}
-			if tmp != 2 {
-				Print("You can't launch it here.", Newline)
-				return true
-			}
-			return true
-		}
-		if (ActVerb.Norm == "drop" && DirObj.Has(FlgWeapon)) ||
-			(ActVerb.Norm == "put" && DirObj.Has(FlgWeapon) && IndirObj == &InflatedBoat) ||
-			((ActVerb.Norm == "attack" || ActVerb.Norm == "mung") && IndirObj != nil && IndirObj.Has(FlgWeapon)) {
-			RemoveCarefully(&InflatedBoat)
-			PuncturedBoat.MoveTo(Here)
-			Rob(&InflatedBoat, Here, 0)
-			Winner.MoveTo(Here)
-			Print("It seems that the ", NoNewline)
-			if ActVerb.Norm == "drop" || ActVerb.Norm == "put" {
-				PrintObject(DirObj)
-			} else {
-				PrintObject(IndirObj)
-			}
-			Print(" didn't agree with the boat, as evidenced by the loud hissing noise issuing therefrom. With a pathetic sputter, the boat deflates, leaving you without.", Newline)
-			if Here.Has(FlgNonLand) {
-				NewLine()
-				if Here == &Reservoir || Here == &InStream {
-					JigsUp("Another pathetic sputter, this time from you, heralds your drowning.", false)
-				} else {
-					JigsUp("In other words, fighting the fierce currents of the Frigid River. You manage to hold your own for a bit, but then you are carried over a waterfall and into some nasty rocks. Ouch!", false)
-				}
-			}
-			return true
-		}
-		return false
-	}
-	if ActVerb.Norm == "board" {
-		if Sceptre.IsIn(Winner) || Knife.IsIn(Winner) || Sword.IsIn(Winner) || RustyKnife.IsIn(Winner) || Axe.IsIn(Winner) || Stiletto.IsIn(Winner) {
-			Print("Oops! Something sharp seems to have slipped and punctured the boat. The boat deflates to the sounds of hissing, sputtering, and cursing.", Newline)
-			RemoveCarefully(&InflatedBoat)
-			PuncturedBoat.MoveTo(Here)
-			ThisIsIt(&PuncturedBoat)
-			return true
-		}
-		return false
-	}
-	if ActVerb.Norm == "inflate" || ActVerb.Norm == "fill" {
-		Print("Inflating it further would probably burst it.", Newline)
-		return true
-	}
-	if ActVerb.Norm == "deflate" {
-		if Winner.Location() == &InflatedBoat {
-			Print("You can't deflate the boat while you're in it.", Newline)
-			return true
-		}
-		if !InflatedBoat.IsIn(Here) {
-			Print("The boat must be on the ground to be deflated.", Newline)
-			return true
-		}
-		Print("The boat deflates.", Newline)
-		Deflate = true
-		RemoveCarefully(&InflatedBoat)
-		InflatableBoat.MoveTo(Here)
-		ThisIsIt(&InflatableBoat)
-		return true
 	}
 	return false
 }
@@ -3801,243 +2753,9 @@ func DeadFunction(arg ActArg) bool {
 	return true
 }
 
-func Awaken(o *Object) bool {
-	s := o.Strength
-	if s < 0 {
-		o.Strength = -s
-		if o.Action != nil {
-			o.Action(ActArg(FConscious))
-		}
-	}
-	return true
-}
-
 // ================================================================
 // INTERRUPT ROUTINES
 // ================================================================
-
-func IFight() bool {
-	if Dead {
-		return false
-	}
-	fightQ := false
-	numVillains := len(Villains)
-	for cnt := 0; cnt < numVillains; cnt++ {
-		oo := Villains[cnt]
-		o := oo.Villain
-		if o.IsIn(Here) && !o.Has(FlgInvis) {
-			if o == &Thief && ThiefEngrossed {
-				ThiefEngrossed = false
-			} else if o.Strength < 0 {
-				p := oo.Prob
-				if p != 0 && Prob(p, false) {
-					oo.Prob = 0
-					Awaken(o)
-				} else {
-					oo.Prob = p + 25
-				}
-			} else if o.Has(FlgFight) || (o.Action != nil && o.Action(ActArg(FFirst))) {
-				fightQ = true
-			}
-		} else {
-			if o.Has(FlgFight) {
-				if o.Action != nil {
-					o.Action(ActArg(FBusy))
-				}
-			}
-			if o == &Thief {
-				ThiefEngrossed = false
-			}
-			Winner.Take(FlgStagg)
-			o.Take(FlgStagg)
-			o.Take(FlgFight)
-			Awaken(o)
-		}
-	}
-	if !fightQ {
-		return false
-	}
-	return DoFight(numVillains)
-}
-
-func ISword() bool {
-	if Sword.IsIn(&Adventurer) {
-		ng := 0
-		if Infested(Here) {
-			ng = 2
-		} else {
-			// Check adjacent rooms for monsters
-			dirs := []*DirProps{
-				&Here.North, &Here.South, &Here.East, &Here.West,
-				&Here.NorthEast, &Here.NorthWest, &Here.SouthEast, &Here.SouthWest,
-				&Here.Up, &Here.Down, &Here.Into, &Here.Out,
-			}
-			for _, dp := range dirs {
-				if dp.IsSet() && dp.RExit != nil {
-					if Infested(dp.RExit) {
-						ng = 1
-						break
-					}
-				}
-			}
-		}
-		g := Sword.TValue
-		if ng == g {
-			return false
-		}
-		if ng == 2 {
-			Print("Your sword has begun to glow very brightly.", Newline)
-		} else if ng == 1 {
-			Print("Your sword is glowing with a faint blue glow.", Newline)
-		} else if ng == 0 {
-			Print("Your sword is no longer glowing.", Newline)
-		}
-		Sword.TValue = ng
-		return true
-	}
-	// Sword not held - disable the interrupt
-	QueueInt(ISword, false).Run = false
-	return false
-}
-
-func IThief() bool {
-	rm := Thief.Location()
-	hereQ := !Thief.Has(FlgInvis)
-	if hereQ {
-		rm = Thief.Location()
-	}
-	flg := false
-	once := false
-	for {
-		if rm == &TreasureRoom && rm != Here {
-			if hereQ {
-				HackTreasures()
-				hereQ = false
-			}
-			DepositBooty(&TreasureRoom)
-		} else if rm == Here && !Here.Has(FlgOn) && !Troll.IsIn(Here) {
-			if ThiefVsAdventurer(hereQ) {
-				return true
-			}
-			if Thief.Has(FlgInvis) {
-				hereQ = false
-			}
-		} else {
-			if Thief.IsIn(rm) && !Thief.Has(FlgInvis) {
-				// Leave if victim left
-				Thief.Give(FlgInvis)
-				hereQ = false
-			}
-			if rm != nil && rm.Has(FlgTouch) {
-				Rob(rm, &Thief, 75)
-				if rm.Has(FlgMaze) && Here.Has(FlgMaze) {
-					flg = RobMaze(rm)
-				} else {
-					flg = StealJunk(rm)
-				}
-			}
-		}
-		if !once && !hereQ {
-			once = true
-			// Move to next room
-			RecoverStiletto()
-			found := false
-			for _, r := range Rooms.Children {
-				if !r.Has(FlgSacred) && r.Has(FlgRLand) {
-					Thief.MoveTo(r)
-					Thief.Take(FlgFight)
-					Thief.Give(FlgInvis)
-					ThiefHere = false
-					rm = r
-					found = true
-					break
-				}
-			}
-			if !found {
-				break
-			}
-			continue
-		}
-		break
-	}
-	if rm != &TreasureRoom {
-		DropJunk(rm)
-	}
-	return flg
-}
-
-func ThiefVsAdventurer(hereQ bool) bool {
-	if !Dead && Here == &TreasureRoom {
-		return false
-	}
-	if !ThiefHere {
-		if !Dead && !hereQ && Prob(30, false) {
-			if Stiletto.IsIn(&Thief) {
-				Thief.Take(FlgInvis)
-				Print("Someone carrying a large bag is casually leaning against one of the walls here. He does not speak, but it is clear from his aspect that the bag will be taken only over his dead body.", Newline)
-				ThiefHere = true
-				return true
-			}
-		}
-		if hereQ && Thief.Has(FlgFight) && !Winning(&Thief) {
-			Print("Your opponent, determining discretion to be the better part of valor, decides to terminate this little contretemps. With a rueful nod of his head, he steps backward into the gloom and disappears.", Newline)
-			Thief.Give(FlgInvis)
-			Thief.Take(FlgFight)
-			RecoverStiletto()
-			return true
-		}
-	}
-	return false
-}
-
-// DropJunk - thief drops valueless items from his bag
-func DropJunk(rm *Object) bool {
-	if rm == nil {
-		return false
-	}
-	flg := false
-	for _, x := range Thief.Children {
-		if x == &Stiletto || x == &LargeBag {
-			continue
-		}
-		if x.TValue == 0 && Prob(30, true) {
-			x.Take(FlgInvis)
-			x.MoveTo(rm)
-			if !flg && rm == Here {
-				Print("The robber, rummaging through his bag, dropped a few items he found valueless.", Newline)
-				flg = true
-			}
-		}
-	}
-	return flg
-}
-
-// StealJunk - thief steals worthless items from a room
-func StealJunk(rm *Object) bool {
-	if rm == nil {
-		return false
-	}
-	for _, x := range rm.Children {
-		if x.TValue == 0 && x.Has(FlgTake) && !x.Has(FlgSacred) && !x.Has(FlgInvis) {
-			if x == &Stiletto || Prob(10, true) {
-				x.MoveTo(&Thief)
-				x.Give(FlgTouch)
-				x.Give(FlgInvis)
-				if x == &Rope {
-					DomeFlag = false
-				}
-				if rm == Here {
-					Print("You suddenly notice that the ", NoNewline)
-					PrintObject(x)
-					Print(" vanished.", Newline)
-					return true
-				}
-				return false
-			}
-		}
-	}
-	return false
-}
 
 func ICandles() bool {
 	Candles.Give(FlgTouch)
@@ -4184,487 +2902,6 @@ func IForestRandom() bool {
 	if Prob(15, false) {
 		Print("You hear in the distance the chirping of a song bird.", Newline)
 	}
-	return true
-}
-
-func IRfill() bool {
-	Reservoir.Give(FlgNonLand)
-	Reservoir.Take(FlgRLand)
-	DeepCanyon.Take(FlgTouch)
-	LoudRoom.Take(FlgTouch)
-	if Trunk.IsIn(&Reservoir) {
-		Trunk.Give(FlgInvis)
-	}
-	LowTide = false
-	if Here == &Reservoir {
-		if Winner.Location().Has(FlgVeh) {
-			Print("The boat lifts gently out of the mud and is now floating on the reservoir.", Newline)
-		} else {
-			JigsUp("You are lifted up by the rising river! You try to swim, but the currents are too strong. You come closer, closer to the awesome structure of Flood Control Dam #3. The dam beckons to you. The roar of the water nearly deafens you, but you remain conscious as you tumble over the dam toward your certain doom among the rocks at its base.", false)
-		}
-	} else if Here == &DeepCanyon {
-		Print("A sound, like that of flowing water, starts to come from below.", Newline)
-	} else if Here == &LoudRoom {
-		Print("All of a sudden, an alarmingly loud roaring sound fills the room. Filled with fear, you scramble away.", Newline)
-		dest := LoudRuns[rand.Intn(len(LoudRuns))]
-		Goto(dest, true)
-	} else if Here == &ReservoirNorth || Here == &ReservoirSouth {
-		Print("You notice that the water level has risen to the point that it is impossible to cross.", Newline)
-	}
-	return true
-}
-
-func IRempty() bool {
-	Reservoir.Give(FlgRLand)
-	Reservoir.Take(FlgNonLand)
-	DeepCanyon.Take(FlgTouch)
-	LoudRoom.Take(FlgTouch)
-	Trunk.Take(FlgInvis)
-	LowTide = true
-	if Here == &Reservoir && Winner.Location().Has(FlgVeh) {
-		Print("The water level has dropped to the point at which the boat can no longer stay afloat. It sinks into the mud.", Newline)
-	} else if Here == &DeepCanyon {
-		Print("The roar of rushing water is quieter now.", Newline)
-	} else if Here == &ReservoirNorth || Here == &ReservoirSouth {
-		Print("The water level is now quite low here and you could easily cross over to the other side.", Newline)
-	}
-	return true
-}
-
-func IMaintRoom() bool {
-	hereQ := Here == &MaintenanceRoom
-	if hereQ {
-		Print("The water level here is now ", NoNewline)
-		idx := WaterLevel / 2
-		if idx >= 0 && idx < len(Drownings) {
-			Print(Drownings[idx], NoNewline)
-		}
-		NewLine()
-	}
-	WaterLevel++
-	if WaterLevel >= 14 {
-		MungRoom(&MaintenanceRoom, "The room is full of water and cannot be entered.")
-		QueueInt(IMaintRoom, false).Run = false
-		if hereQ {
-			JigsUp("I'm afraid you have done drowned yourself.", false)
-		}
-	} else if InflatedBoat.IsIn(Winner) && (Here == &MaintenanceRoom || Here == &DamRoom || Here == &DamLobby) {
-		JigsUp("The rising water carries the boat over the dam, down the river, and over the falls. Tsk, tsk.", false)
-	}
-	return true
-}
-
-func IRiver() bool {
-	if Here != &River1 && Here != &River2 && Here != &River3 && Here != &River4 && Here != &River5 {
-		QueueInt(IRiver, false).Run = false
-		return false
-	}
-	rm := Lkp(Here, RiverNext)
-	if rm != nil {
-		Print("The flow of the river carries you downstream.", Newline)
-		NewLine()
-		Goto(rm, true)
-		// ZIL: <ENABLE <QUEUE I-RIVER <LKP ,HERE ,RIVER-SPEEDS>>>
-		// After Goto, HERE is the new room. Use its speed for the next leg.
-		if spd, ok := RiverSpeedMap[Here]; ok {
-			Queue(IRiver, spd).Run = true
-		}
-		return true
-	}
-	JigsUp("Unfortunately, the magic boat doesn't provide protection from the rocks and boulders one meets at the bottom of waterfalls. Including this one.", false)
-	return true
-}
-
-// Remark prints a combat message, substituting weapon/defender names for markers
-func Remark(msg MeleeMsg, defender, weapon *Object) {
-	for _, part := range msg {
-		switch part.Marker {
-		case FWep:
-			if weapon != nil {
-				PrintObject(weapon)
-			}
-		case FDef:
-			if defender != nil {
-				PrintObject(defender)
-			}
-		default:
-			Print(part.Text, NoNewline)
-		}
-	}
-	NewLine()
-}
-
-// RandomMeleeMsg picks a random message from a MeleeSet
-func RandomMeleeMsg(set MeleeSet) MeleeMsg {
-	if len(set) == 0 {
-		return nil
-	}
-	return set[rand.Intn(len(set))]
-}
-
-// VillainStrength calculates a villain's effective combat strength
-func VillainStrength(oo *VillainEntry) int {
-	od := oo.Villain.Strength
-	if od >= 0 {
-		if oo.Villain == &Thief && ThiefEngrossed {
-			if od > 2 {
-				od = 2
-			}
-			ThiefEngrossed = false
-		}
-		if IndirObj != nil && IndirObj.Has(FlgWeapon) && oo.Best == IndirObj {
-			tmp := od - oo.BestAdv
-			if tmp < 1 {
-				tmp = 1
-			}
-			od = tmp
-		}
-	}
-	return od
-}
-
-// VillainResult applies the combat result to a villain
-func VillainResult(villain *Object, def int, res BlowRes) BlowRes {
-	villain.Strength = def
-	if def == 0 {
-		villain.Take(FlgFight)
-		Print("Almost as soon as the ", NoNewline)
-		PrintObject(villain)
-		Print(" breathes his last breath, a cloud of sinister black fog envelops him, and when the fog lifts, the carcass has disappeared.", Newline)
-		RemoveCarefully(villain)
-		if villain.Action != nil {
-			villain.Action(ActArg(FDead))
-		}
-		return res
-	}
-	if res == BlowUncon {
-		if villain.Action != nil {
-			villain.Action(ActArg(FUnconscious))
-		}
-		return res
-	}
-	return res
-}
-
-// WinnerResult applies the combat result to the player
-func WinnerResult(def int, res BlowRes, od int) {
-	if def == 0 {
-		Winner.Strength = -10000
-	} else {
-		Winner.Strength = def - od
-	}
-	if def-od < 0 {
-		Queue(ICure, CureWait).Run = true
-	}
-	if FightStrength(true) <= 0 {
-		Winner.Strength = 1 - FightStrength(false)
-		JigsUp("It appears that that last blow was too much for you. I'm afraid you are dead.", false)
-	}
-}
-
-// VillainBlow processes a villain's attack on the player
-func VillainBlow(oo *VillainEntry, out bool) BlowRes {
-	villain := oo.Villain
-	remarks := oo.Msgs
-
-	Winner.Take(FlgStagg)
-	if villain.Has(FlgStagg) {
-		Print("The ", NoNewline)
-		PrintObject(villain)
-		Print(" slowly regains his feet.", Newline)
-		villain.Take(FlgStagg)
-		return BlowMissed
-	}
-	att := VillainStrength(oo)
-	oa := att
-	def := FightStrength(true)
-	if def <= 0 {
-		return BlowMissed
-	}
-	od := FightStrength(false)
-	dweapon := FindWeapon(Winner)
-	var res BlowRes
-	if def < 0 {
-		res = BlowKill
-	} else {
-		if def == 1 {
-			if att > 2 {
-				att = 3
-			}
-			idx := att - 1
-			if idx < 0 {
-				idx = 0
-			}
-			if idx >= len(Def1Res) {
-				idx = len(Def1Res) - 1
-			}
-			tbl := Def1Res[idx]
-			_ = tbl
-			// Lookup from the actual Def arrays
-			switch {
-			case att <= 1:
-				res = Def1[rand.Intn(len(Def1))]
-			case att == 2:
-				if int(Def1Res[1]) < len(Def1) {
-					start := int(Def1Res[1])
-					_ = start
-				}
-				res = Def1[rand.Intn(len(Def1))]
-			default:
-				res = Def1[rand.Intn(len(Def1))]
-			}
-		} else if def == 2 {
-			if att > 3 {
-				att = 4
-			}
-			switch {
-			case att <= 1:
-				res = Def2A[rand.Intn(len(Def2A))]
-			default:
-				res = Def2B[rand.Intn(len(Def2B))]
-			}
-		} else {
-			att = att - def
-			if att < -1 {
-				att = -2
-			}
-			if att > 1 {
-				att = 2
-			}
-			idx := att + 2
-			if idx < 0 {
-				idx = 0
-			}
-			if idx > 4 {
-				idx = 4
-			}
-			switch {
-			case idx <= 1:
-				res = Def3A[rand.Intn(len(Def3A))]
-			case idx == 2:
-				res = Def3B[rand.Intn(len(Def3B))]
-			default:
-				res = Def3C[rand.Intn(len(Def3C))]
-			}
-		}
-		if out {
-			if res == BlowStag {
-				res = BlowHesitate
-			} else {
-				res = BlowSitDuck
-			}
-		}
-		if res == BlowStag && dweapon != nil && Prob(25, true) {
-			res = BlowLoseWpn
-		}
-		msg := RandomMeleeMsg((*remarks)[res-1])
-		if msg != nil {
-			Remark(msg, Winner, dweapon)
-		}
-	}
-	_ = oa
-	// Apply results
-	switch {
-	case res == BlowMissed || res == BlowHesitate:
-		// Nothing
-	case res == BlowUncon:
-		// Nothing extra
-	case res == BlowKill || res == BlowSitDuck:
-		def = 0
-	case res == BlowLightWnd:
-		def--
-		if def < 0 {
-			def = 0
-		}
-		if LoadAllowed > 50 {
-			LoadAllowed -= 10
-		}
-	case res == BlowHeavyWnd:
-		def -= 2
-		if def < 0 {
-			def = 0
-		}
-		if LoadAllowed > 50 {
-			LoadAllowed -= 20
-		}
-	case res == BlowStag:
-		Winner.Give(FlgStagg)
-	default:
-		// BlowLoseWpn
-		if dweapon != nil {
-			dweapon.MoveTo(Here)
-			nweapon := FindWeapon(Winner)
-			if nweapon != nil {
-				Print("Fortunately, you still have a ", NoNewline)
-				PrintObject(nweapon)
-				Print(".", Newline)
-			}
-		}
-	}
-	WinnerResult(def, res, od)
-	return res
-}
-
-// DoFight runs the villain combat loop
-func DoFight(numVillains int) bool {
-	for {
-		cnt := 0
-		var res BlowRes
-		out := false
-		for cnt < numVillains {
-			oo := Villains[cnt]
-			cnt++
-			o := oo.Villain
-			if !o.Has(FlgFight) {
-				continue
-			}
-			if o.Action != nil && o.Action(ActArg(FBusy)) {
-				continue
-			}
-			res = VillainBlow(oo, out)
-			if res == BlowUnk {
-				return false
-			}
-			if res == BlowUncon {
-				out = true
-			}
-		}
-		if res != BlowUnk {
-			if !out {
-				return true
-			}
-			// unconscious rounds
-			if out {
-				return true
-			}
-		}
-		return true
-	}
-}
-
-func HeroBlow() bool {
-	if DirObj == nil {
-		return false
-	}
-	// Find the villain entry for the target
-	var oo *VillainEntry
-	for _, v := range Villains {
-		if v.Villain == DirObj {
-			oo = v
-			break
-		}
-	}
-	DirObj.Give(FlgFight)
-	if Winner.Has(FlgStagg) {
-		Print("You are still recovering from that last blow, so your attack is ineffective.", Newline)
-		Winner.Take(FlgStagg)
-		return true
-	}
-	att := FightStrength(true)
-	if att < 1 {
-		att = 1
-	}
-	oa := att
-	_ = oa
-	villain := DirObj
-	if oo == nil {
-		return false
-	}
-	od := VillainStrength(oo)
-	def := od
-	if def == 0 {
-		if DirObj == Winner {
-			JigsUp("Well, you really did it that time. Is suicide painless?", false)
-			return true
-		}
-		Print("Attacking the ", NoNewline)
-		PrintObject(villain)
-		Print(" is pointless.", Newline)
-		return true
-	}
-	dweapon := FindWeapon(villain)
-	var res BlowRes
-	if dweapon == nil || def < 0 {
-		Print("The ", NoNewline)
-		if def < 0 {
-			Print("unconscious", NoNewline)
-		} else {
-			Print("unarmed", NoNewline)
-		}
-		Print(" ", NoNewline)
-		PrintObject(villain)
-		Print(" cannot defend himself: He dies.", Newline)
-		res = BlowKill
-	} else {
-		if def == 1 {
-			if att > 2 {
-				att = 3
-			}
-			res = Def1[rand.Intn(len(Def1))]
-		} else if def == 2 {
-			if att > 3 {
-				att = 4
-			}
-			res = Def2B[rand.Intn(len(Def2B))]
-		} else {
-			att = att - def
-			if att < -1 {
-				att = -2
-			}
-			if att > 1 {
-				att = 2
-			}
-			idx := att + 2
-			if idx < 0 {
-				idx = 0
-			}
-			switch {
-			case idx <= 1:
-				res = Def3A[rand.Intn(len(Def3A))]
-			case idx == 2:
-				res = Def3B[rand.Intn(len(Def3B))]
-			default:
-				res = Def3C[rand.Intn(len(Def3C))]
-			}
-		}
-		if res == BlowStag && dweapon != nil && Prob(25, false) {
-			res = BlowLoseWpn
-		}
-		msg := RandomMeleeMsg(HeroMelee[res-1])
-		if msg != nil {
-			Remark(msg, DirObj, IndirObj)
-		}
-	}
-	// Apply hero blow results
-	switch {
-	case res == BlowMissed || res == BlowHesitate:
-		// Nothing
-	case res == BlowUncon:
-		def = -def
-	case res == BlowKill || res == BlowSitDuck:
-		def = 0
-	case res == BlowLightWnd:
-		def--
-		if def < 0 {
-			def = 0
-		}
-	case res == BlowHeavyWnd:
-		def -= 2
-		if def < 0 {
-			def = 0
-		}
-	case res == BlowStag:
-		DirObj.Give(FlgStagg)
-	default:
-		// BlowLoseWpn
-		if dweapon != nil {
-			dweapon.Take(FlgNoDesc)
-			dweapon.Give(FlgWeapon)
-			dweapon.MoveTo(Here)
-			ThisIsIt(dweapon)
-		}
-	}
-	VillainResult(DirObj, def, res)
 	return true
 }
